changed CSV to URL, "RMPSE" to "RMSE" for training set, and Scaled.
I changed the import code to a URL link bc the rubric asks that no additional materials are needed to run our project.  I also changed the language in the K-NN section so that "RMPSE" is changed to "RMSE" when speaking about the training data set.

The biggest change in our project is the addition of scaling, which was previously missing. This is the tricky thing because I know you're not supposed to scale one hot and ordinal encoding, but I don't know if coding the StandardScaler to be only on "age" is correct either...Help.